15th lesson, improving css and drobbing into smaller components
