+++ src/App.js
@@ -1,8 +1,8 @@
 import logo from './logo.svg';
 import './App.css';
-import Header from './components/Header';
-import Profile from './components/Profile';
-import Nav from './components/Nav';
+import Header from './components/Header/Header';
+import Profile from './components/Profile/Profile';
+import Nav from './components/Navbar/Nav';
 
 function App() {
   return (
