+++ src/components/Header/Header.module.css
@@ -4,6 +4,8 @@
   border-radius: 40px;
   box-shadow: 0 0 5px 1px;
   margin-top: 10px;
+  position: sticky;
+  top: 10px;
 }
 .header img {
   width: 70px;
