+++ src/components/Profile/MyPosts/MyPosts.jsx
@@ -1,0 +1,25 @@
+import React from "react";
+import s from "./MyPosts.module.css";
+import Post from "./Post/Post";
+
+const MyPosts = () => {
+  return (
+    <div>
+      My posts
+      <div>
+        <textarea></textarea>
+        <button>Add button</button>
+      </div>
+      <div>New posts</div>
+      <div className={s.posts}>
+        <Post/>
+        <Post/>
+        <Post/>
+        <Post/>
+        <Post/>
+        <Post/>
+      </div>
+    </div>
+  );
+};
+export default MyPosts;
