+++ src/components/Profile/MyPosts/MyPosts.module.css
@@ -1,0 +1,7 @@
+.item {
+    color: greenyellow;
+}
+.item img {
+  width: 50px;
+  height: 50px;
+}
