+++ src/components/Profile/MyPosts/Post/Post.module.css
@@ -1,0 +1,8 @@
+.item {
+    color: greenyellow;
+}
+.item img {
+  width: 50px;
+  height: 50px;
+  border-radius: 30px;
+}
