+++ src/components/Profile/Profile.module.css
@@ -6,6 +6,3 @@
   height: 300px;
   width: 900px;
 }
-.item {
-    color: greenyellow;
-}
